+++ tests/std.rs
@@ -4,6 +4,12 @@ macro_rules! search_test {
         assert_eq!(r, $result);
     };
 }
+macro_rules! rsearch_test {
+    ($haystack:expr, $needle:expr, $result:expr) => {
+        let r = $haystack.rfind($needle);
+        assert_eq!(r, $result);
+    };
+}
 
 #[cfg(test)]
 mod std_str_str {
@@ -194,6 +200,53 @@ mod std_string_string {
 }
 
 #[cfg(test)]
+mod std_str_str_rev {
+    #[test]
+    fn test_empty_needle() {
+        rsearch_test!("", "", Some(0)); // ***
+        rsearch_test!("1", "", Some(1)); // ***
+        let haystack = "111 a 111b";
+        rsearch_test!(haystack, "", Some(10)); // ***
+    }
+    #[test]
+    fn test_not_found() {
+        let haystack = "111 a 111b";
+        rsearch_test!(haystack, "xxx", None);
+        rsearch_test!(haystack, "12b", None);
+        rsearch_test!(haystack, "a31", None);
+    }
+    #[test]
+    fn test_perfect_matching() {
+        let haystack = "111 a 111b";
+        let needle = "111 a 111b";
+        rsearch_test!(haystack, needle, Some(0));
+    }
+    #[test]
+    fn test_same_size() {
+        let haystack = "111 a 111b";
+        let needle = "111 a 1 1b";
+        rsearch_test!(haystack, needle, None);
+    }
+    #[test]
+    fn test_large_needle() {
+        let haystack = "111 a 111b";
+        let needle = "111 a 111bZ";
+        rsearch_test!(haystack, needle, None);
+    }
+    #[test]
+    fn test_last_match() {
+        let haystack = "111 a 111b";
+        rsearch_test!(haystack, "b", Some(9));
+    }
+    #[test]
+    fn test_small_needle() {
+        let haystack = "111 a 111b";
+        rsearch_test!(haystack, "a", Some(4));
+        rsearch_test!(haystack, "a 111", Some(4));
+    }
+}
+
+#[cfg(test)]
 mod std_match_indices {
     #[test]
     fn test_match_indices() {
@@ -206,6 +259,23 @@ mod std_match_indices {
         assert_eq!(m.next(), Some((6, "1")));
         assert_eq!(m.next(), Some((7, "1")));
         assert_eq!(m.next(), Some((8, "1")));
+        assert_eq!(m.next(), None);
+    }
+}
+
+#[cfg(test)]
+mod std_rmatch_indices {
+    #[test]
+    fn test_rmatch_indices() {
+        let haystack = "111 a 111b";
+        let needle = "1";
+        let mut m = haystack.rmatch_indices(needle);
+        assert_eq!(m.next(), Some((8, "1")));
+        assert_eq!(m.next(), Some((7, "1")));
+        assert_eq!(m.next(), Some((6, "1")));
+        assert_eq!(m.next(), Some((2, "1")));
+        assert_eq!(m.next(), Some((1, "1")));
+        assert_eq!(m.next(), Some((0, "1")));
         assert_eq!(m.next(), None);
     }
 }