+++ CHANGELOG.md
@@ -2,6 +2,15 @@ TBD
 ===
 Unreleased changes. Release notes have not yet been written.
 
+0.1.5 (2021-03-20)
+=====
+
+* add Search::rsearch()
+* add Search::rsearch_indices()
+* add string_rsearch()
+* add string_rsearch_indices()
+* add more docs
+
 0.1.4 (2021-03-19)
 =====